pyproject.toml
[project]
name = "child-allowance-tracker"
version = "0.1.0"
description = "Child allowance tracking app"
readme = "README.md"
requires-python = ">=3.12"  # Changed from 3.13
dependencies = [
    "flask>=2.3.0",
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.95.0",
    "gspread>=5.10.0",
    "python-dotenv>=1.0.0",
    "awsgi>=0.2.7",
    "boto3>=1.28.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]